refactor: 코드래빗 피드백 수정

+++ server/src/main/java/com/server/inheritance/dto/InheritanceSummaryDto.java
@@ -17,6 +17,6 @@ public class InheritanceSummaryDto {
   private Long id;
   private String username;
   private BigDecimal percent;
-  private Integer amt;
+  private Long amt;
 
 }
